requirements.txt
flask-smorest
mongoengine
dateutils
unidecode
flask-cors
numpy
voxpopuli
tqdm
marshmallow_enum
marshmallow
dataclasses
blinker
python-dateutil
python-dotenv
-e git://github.com/loult-elte-fwere/cookie-factory.git#egg=cookie-factory
python-magic
pydub